pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.14"

dependencies = [
    # LiveKit core
    "livekit",
    "livekit-api",
    "livekit-protocol",

    # LiveKit Agents
    "livekit-agents[silero,turn-detector]~=1.3",  #CUDA, comentar si no GPU nvidia

    # LiveKit Plugins
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai",
    "livekit-plugins-deepgram",
    "livekit-plugins-elevenlabs",
    "livekit-plugins-silero",                     #CUDA, comentar si no GPU nvidia

    # Speech / TTS / STT
    "openai-whisper",                             #CUDA, comentar si no GPU nvidia
    "piper-tts",                                  #Se cambia piper-tts por gtts
    "gtts",

    # Web / API
    "flask",
    "flask-cors",
    "requests",

    # Config
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
